Fixed: activity_login.xml activity_register.xml activity_main.xml
We also restructured the files and used MVC structure method for more easy and convinent way to get data from the backend to the frontend. And it made us understand better the communication between the backend and the front end respectivly.

Added:
activity_update_profile.xml - for the user profile picture.

UserApi.java - Models for database for users.

ApiClient.java for creating and configuring a single instance of Retrofit that your whole app can use to make API calls to your backend (Spring Boot server) + using singleton design pattern.

LoginRequest.java - The main role of LoginRequest is to represent the user's login credentials when you're sending a login request from Android to the backend via Retrofit.

RegisterActivity.java - is a data model class used to send user information from the Android app to the backend when registering a new user.

Updated:

AndroidManifest.xml to show activities as the user requests.

And updated the entire project with a new H2 database which saves the registered users locally.